[CELEBORN-INNER] Move Utils to CheckUtils
