+++ client-flink/common/src/main/java/org/apache/celeborn/plugin/flink/buffer/PartitionSortedBuffer.java
@@ -17,9 +17,9 @@
 
 package org.apache.celeborn.plugin.flink.buffer;
 
-import static org.apache.celeborn.plugin.flink.utils.Utils.checkArgument;
-import static org.apache.celeborn.plugin.flink.utils.Utils.checkNotNull;
-import static org.apache.celeborn.plugin.flink.utils.Utils.checkState;
+import static org.apache.celeborn.common.util.CheckUtils.checkArgument;
+import static org.apache.celeborn.common.util.CheckUtils.checkNotNull;
+import static org.apache.celeborn.common.util.CheckUtils.checkState;
 import static org.apache.flink.runtime.io.network.buffer.Buffer.DataType;
 
 import java.io.IOException;
